Move validation logic to AppRegistryService
