+++ spring-cloud-dataflow-server-core/src/main/java/org/springframework/cloud/dataflow/server/controller/AppRegistryController.java
@@ -60,7 +60,6 @@ import org.springframework.hateoas.PagedResources;
 import org.springframework.hateoas.mvc.ResourceAssemblerSupport;
 import org.springframework.http.HttpStatus;
 import org.springframework.util.CollectionUtils;
-import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -204,19 +203,7 @@ public class AppRegistryController {
 			@RequestParam("uri") String uri, @RequestParam(name = "metadata-uri", required = false) String metadataUri,
 			@RequestParam(value = "force", defaultValue = "false") boolean force) {
 
-		// compare existing default app and deny registration if base path of uri's
-		// don't match as registering app with different version can only differ by version
-		AppRegistration defaultApp = appRegistryService.getDefaultApp(name, type);
-		if  (defaultApp != null && StringUtils.hasText(version)) {
-			String defaultAppUri = defaultApp.getUri().toString();
-			String defaultAppUriNoVersion = removeLastMatch(defaultAppUri, defaultApp.getVersion());
-			String newAppUriNoVersion = removeLastMatch(uri, version);
-			if (!ObjectUtils.nullSafeEquals(defaultAppUriNoVersion, newAppUriNoVersion)) {
-				throw new IllegalArgumentException("Existing default application [" + defaultAppUri
-						+ "] can only differ by a version but is [" + uri + "]");
-			}
-		}
-
+		appRegistryService.validate(appRegistryService.getDefaultApp(name, type), uri, version);
 		AppRegistration previous = appRegistryService.find(name, type, version);
 		if (!force && previous != null) {
 			throw new AppAlreadyRegisteredException(previous);
@@ -229,14 +216,6 @@ public class AppRegistryController {
 		catch (URISyntaxException e) {
 			throw new IllegalArgumentException(e);
 		}
-	}
-
-	private static String removeLastMatch(String original, String match) {
-		StringBuilder builder = new StringBuilder();
-		int start = original.lastIndexOf(match);
-		builder.append(original.substring(0, start));
-		builder.append(original.substring(start + match.length()));
-		return builder.toString();
 	}
 
 	@Deprecated
